elf: fix libbpf integration tests
The integration tests are currently broken. When running them in parallel,
the file variable is captured by the closure in t.Run, and so we only
ever try to load a single file.

Fix this by taking an explicit copy of the filename, and by turnin off
parallel tests. This is turn makes the tests fail, since there is plenty
of libbpf stuff that we don't understand.